requirements.txt
rsa==4.7.2
pickleshare==0.7.5
paho-mqtt==1.5.1